chore: extract framework code to separate file
